Bugfix for reversing after unloading and header collision
This fixes the reversing behaviour after finishing the unloading procedure of a combine. (Closes #274)
Also the collision with the header and other potential attachements of the combine is now properly checked in the path finding stage. (Fixes #246)
Small fix for using multiple trailers when enabling AutoDrive with an already full trailer and following empty trailers.