Release 1.0.5.9
Work-in-progress
Please be aware that I am still developing the chase combine/forrage harvester mode and it will not be working as stable as the standard unloader mode!
Also this update affects parts of the general path finder.
So if there are any issues, please open up new issues. Thank you!