Added option to restrict path planning to field borders
This only applies to vehicles which are already parked in the field.
It doesn't detect the field number or anything right now. So use with caution.
It is active if the vehicle is parked or activated in the field and is thus more than 10m away from the starting location :-) A little workaround until I improve on that function.

@Iwan1803 I hope that stops them from driving into the ditches on NF Marsch ;-) (LG Balu)