New Hud textures
New Hud design. Finally updated from FS17 to FS19 :D
Thanks to OlfDE for his input to creating the new design.
If you feel like thanking him for his work. Here is a link: https://streamlabs.com/olfde/tip
He was also the creator of the original HUD which was perfectly suited for FS17. So he is the brains behind the HUD Design since the beginning :-)

Also increased the smoothing of paths on the field and included some learning from the 'Claas DLC' incompatabilities to ensure this won't happen again.