Bugfixes and improvements
If CP harvester has finished it's route, AD will drive to unload #729
Bugfix in pathfinder. The start on field detection was accidentally run on the combine location and not on the unload driver's position.
So if restrict to field is enabled, the pathfinder should now work better and not choose a path through the fruit.

Fixed driver stopping after unloading. #626