Start Courseplay fixed
The Courseplay team broke the courseplay:start() interface, that
can't be used anymore to start the CP driver.

This fix addresses that issue (see:
https://github.com/Courseplay/courseplay/issues/7062)

The fix uses the two explicit interfaces to start/stop the
Courseplay driver. These will be part of the Courseplay version
> 6.03.00053.

Kept the old start/stop call for backward compatibility.

Moved the start code to ExternalInterface.lua.

It seems like AutoDrive:StopCP() isn't called for example when AD
runs a refill course for CP and AD is stopped, the AD team may want
to look at that.