Add Feature "lightsHazardsOffField"
When not on the field, allow per-vehicle setting to run with hazard lights on (both turn signals) - Pretty common to see large equipment run down streets like this, at least in the US.  This should override blinkangle correctly off-field, and allow it to still work on-field.  Testing of this to follow.