Add Features "lightsTurnOfTime" and "lightsTurnOffTime"
Allow the user to override the map/seasons/giants default of when headlights should turn on and off.  Pre-poulated off time to be between 00:30 (12:30a) and 10:00 (10a), and on time between 16:00 (4p) and 23:30 (11:30p).  Avoided 00:00 and 24:00 because I have no idea how rounding would affect things, but if wanted, we could internally use 00:01 (1) and 23:59 (1439) and avoid any posible issue.